requirements.txt
# Async HTTP
aiohttp>=3.9.0

# Discord integration
discord.py>=2.3.0

# Binance API
python-binance>=1.0.19
requests>=2.31.0

# IV calculation (fallback)
py-vollib>=1.0.1
scipy>=1.11.0
numpy>=1.24.0

# Configuration & utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Database for ATM IV tracking
aiosqlite>=0.19.0

# Logging
colorlog>=6.7.0
